// frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        PaymentResponse: {
            /**
             * Format: uuid
             * @description Payment ID
             * @example 550e8400-e29b-41d4-a716-446655440000
             */
            id?: string;
            /**
             * Format: uuid
             * @description Invoice ID
             * @example 550e8400-e29b-41d4-a716-446655440000
             */
            invoiceId?: string;
            /**
             * @description Payment amount
             * @example 500
             */
            amount?: number;
            /**
             * Format: date-time
             * @description Payment date
             */
            paymentDate?: string;
            /**
             * Format: date-time
             * @description Creation timestamp
             */
            createdAt?: string;
        };
        Pageable: {
            /** Format: int32 */
            page?: number;
            /** Format: int32 */
            size?: number;
            sort?: string[];
        };
        PagePaymentResponse: {
            /** Format: int64 */
            totalElements?: number;
            /** Format: int32 */
            totalPages?: number;
            pageable?: components["schemas"]["PageableObject"];
            /** Format: int32 */
            numberOfElements?: number;
            first?: boolean;
            last?: boolean;
            /** Format: int32 */
            size?: number;
            content?: components["schemas"]["PaymentResponse"][];
            /** Format: int32 */
            number?: number;
            sort?: components["schemas"]["SortObject"];
            empty?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}